Feat (script) converted the scrpit to ts
